+++ runway/lookups/handlers/__init__.py
@@ -1,4 +1,5 @@
 """Import classes."""
+from . import cfn  # noqa: F401
 from . import env  # noqa: F401
 from . import ssm  # noqa: F401
 from . import var  # noqa: F401